refactor: too many lines
