+++ src/renderer/src/components/Product/FilterFields/types.tsx
@@ -1,0 +1,15 @@
+import {
+  ProductSearchFilters,
+  ProductSearchPayloadOperation,
+} from "../Interface";
+
+export interface IFilterProps {
+  title: string;
+  id?: number;
+  handleChange: (
+    value: string | number | boolean,
+    operation?: ProductSearchPayloadOperation
+  ) => void;
+  searchData?: ProductSearchFilters;
+  autoCompleteValues?: (string | number)[];
+}
